Parameterize printf's and usage
As printf substitutes some character sequences (such as `%`), replace all
variables in message strings with `%s` and pass as parameter. This is to
ensure that variables can contain any of these characeters (e.g.,
properly print `%` in path names).

In addition, this will make it easier to provide translations of messages
with `gettext` in the future.