Provide internationalization support via gettext ($"...")
Added Makefile, which will compile .po files in the `po` (to be added in
a future commit) subdirectory.  The Makefile also provides an `install`
target.

A template for .po files will be added in a future commit, but this
requires somewhat decided upon english messages (to be done soon).

By default `pkgix`, if run from a cloned repository, will look for the
`locale` dir in `<path-to-pkgix>/share`. If `pkgix` is installed with
`make install`, then the `TEXTDOMAINDIR` is substituted accordingly.