pkgix: Added `list` options --explicit, --depends, --unrequired
The `list` command can now filter by explictly installed packages,
packages installed as as dependency for another package, unrequires
packages or a combination of them.

In order to differentiate between explitly installed and installed as a
dependency, it as required to to add the `reason` field to the info file
and properly deal with preserving this information on upgrades.

The filter for unrequired packages checks across all listed pkgix-roots,
so that if package A is required by another package B not in the same
pkgix-root, package A won't be listed with this filter.