pkgix: avail can now display packages for remote repositories
If the remote repository provides the file `.avail` in its root, it is
used to display available packages at the remote repository.

The `.avail` file can be easily obtained by running:
`PKGIX_REPO_URLS='<repo>' pkgix avail > <repo>/.avail`
This could be automated when commiting changes to a repository.

As setting IFS in `avail` and only resetting it at the end of the call,
causing an issue with the expansion of $fetch_cmd in `fetch_remote`, it
was necessary to change the way PKGIX_REPO_URLS is expanded in all
places it is used.

As a result, `sanitize_repo_urls` now turns PKGIX_REPO_URLS into an
array; in addition, it is left untouched if it is already an array, as
the PKGIX_REPO_URLS could be set in the RC-file.

Signed-off-by: Marco Elver <me@marcoelver.com>