pkgix: In export_env(), environment check regex should treat pkgix_root as string
Added missing quotes, so the expansion of $pkgix_root in the checks
against PKGIX_ROOT and CPATH do not consider characters in pkgix_root as
part of the regex, but only a string.

This caused a problem where I could have two paths, one being something
like `path-suffix` and the second being `path.suffix` -- the second
would never be added to the environment.

Signed-off-by: Marco Elver <me@marcoelver.com>