pkgix: Properly handle PKGIX_REPO_URLS if set to array, even if empty
PKGIX_REPO_URLS can be set to an empty array from the RC-file, in which
case the user explicitly requested an empty array (repo-urls can still
be specified with `-r`). If this is the case, do not assign this array
the default value.

The default value will be the null string if the default directory does
not exist.

Signed-off-by: Marco Elver <me@marcoelver.com>