pkgix: Add exit_cleanup(), do not call pkgix_upgrade() in subshell
I added exit_cleanup(), so in the future, I can do more cleanup than
just remove the locks on EXIT.

For this to work, I had to change the locking code, which is now keeping
track of which prefix is locked (if at all) using a global variable.

This required changing prefix_upgrade() to not be called in a subshell
anymore, otherwise the variable __lock_held would not be probagated to
the outer-most shell.

Signed-off-by: Marco Elver <me@marcoelver.com>