pkgix: Implemented source_pkg() remote package file caching
source_pkg() now caches downloaded package description files, so that if
they are reused in the same install/upgrade session, they will be loaded
from the cached copy on disk, instead of always refetching.

Repositories are distinguished by hashing the repo URL and creating the
repo-cache path with this hash.

Signed-off-by: Marco Elver <me@marcoelver.com>