pkgix: Install already installed/satisfied packages with -r instead of -f
The -f flag covered a bit too many unsafe conditions to be forced and
ignored, yet installing packages already installed/satisfied by the
system is useful without forcing all other checks.

The -r flag was unused in the case where a package is not yet installed,
so it now has meaning when a package is not yet installed inside a root:
install already installed/satisfied packages.

Signed-off-by: Marco Elver <me@marcoelver.com>