pkgix: Package files can provide optional `provides` field
Package files can now provide an optional provides field, which is a
list of packages which the installed package is providing in addition to
the package name itself.

On installation, for each provides entry, a symlink is made from the
provides package db dir to the real package dir, so that packages
depending on a package provided by an installed package (e.g., gcc-4.4
provides gcc), these dependencies are not installed again.

It is possible to remove a package by one of the names listed in the
provides array; this was necessary, so that if a package which is
already provided is selected to be reinstalled, the real package is
removed first, and then the new package (previously provided) is then
installed.

Signed-off-by: Marco Elver <me@marcoelver.com>