pkgix: Add support for listing dependencies with version information
Packages can now list dependencies in the `depends` field, with
additional version information: <dependency>(>|<|>=|<=|=)<ver>

This helps to ensure that the correct package versions are installed.
This just checks against either already installed packages or checks
against dependencies installed by a target.

This should not be confused with installing packages of a specific
version, by simply providing a package file with the version encoded in
the name. It would be difficult to guess and resolve version issues
without the user specifying which package exactly provides required
version. Such packages would then need to provide an entry in the
`provides` field with the basename of the package.

Signed-off-by: Marco Elver <me@marcoelver.com>