pkgix: Add dry run check against dependency version
When doing a dry run, and an installed package depends on a specific
version of a package (see previous commit), it would always report the
wrong version to be installed, as no package was installed, therefore no
version information is in the db.

When running with the -D flag, if a package is not yet installed, check
against the version as denoted in the dependency's package description
file in the repository.

Factored out the lexicographical comparison to its own function
strcmp_op.

Signed-off-by: Marco Elver <me@marcoelver.com>