pkgix: Topologically sort dependency on upgrade
The db file only tracks the order of packages as they are installed, but
does not take into account packages that are upgraded and pull new
dependencies, or packages which are installed and depend on already
installed packages.

Therefore, it is necessary to topologically sort packages according to
the dependency graph they form.

Signed-off-by: Marco Elver <me@marcoelver.com>