pkgix: Perform version check before installing dependency
Perform version before installing a dependency, and only if this
dependency is already installed.

Checking after was causing problems with metapackages or packages where
the isinstalled method determined the package was installed outside a
prefix, as the database does not track these.

As a result, the version check will now only result in an error if a
dependency package is already installed, but the required version does
not match.

If a package installs a dependency with a wrong version, this will not
result in an error while installing the package, but eventually, either
during upgrades or reinstalls -- these issues should be solved while
creating packages files and maintaining the repository properly (e.g.,
after creating a package, perform an explicit dry install of all
dependencies and the package to test itself, which should report if
there are problems).

Signed-off-by: Marco Elver <me@marcoelver.com>