pkgix: Properly deal with prefix /
If prefix is '/', do not add another slash -- this is a special case, as
usually the path is sanitized to have all trailing slashes removed,
however, in the case of '/' this is not possible.

Note that package files will be provided with an empty $prefix in this
case, only pkgix itself maintains '/' as setting pkgix_prefix to null
could case problems in some parts of the code.

Signed-off-by: Marco Elver <me@marcoelver.com>