pkgix: Better use of Bash, remove IFS_expand, fix `read`
IFS_expand is not necessary as the same can be achieved with the Bash
builtin read: IFS="$ifs" read -ra dest_array <<< "$src_string"
[See http://mywiki.wooledge.org/Arguments]

Change all uses of `read` to `read -r` (Bash best-practice
[http://wiki.bash-hackers.org/commands/builtin/read]).

Set IFS= before `read`, so that whitespace is not removed when reading a
line.

Still an obvious problem: new-line is not allowed in folder or file
names. This is likely not going to be fixed, as it would complicate
storing list of files in 'files' as well.

Signed-off-by: Marco Elver <me@marcoelver.com>