pkgix: Provide support for optional `check`
Package description files can now have an optional `check` function
which can be used to check if the package has been built correctly.

It makes sense to provide a separate function for this, as it would be a
problem if the user would like to ignore a failed check and this test
would be performed in `build` or `installenv`. The `--ignore-check`
option can be used to ignore failed checks.

Signed-off-by: Marco Elver <me@marcoelver.com>