pkgix: Support postremove hook
The postremove hook can be specified in the package description file.

Design decision: when a package is removed, we do not want to rely on
the package description file to exist; I made the decision that I do not
want to copy package description file to the package db, so instead,
just copy the code returned by `declare -f` to a `hooks` file.

This has some drawbacks, such as not being able to use variables
specified in the package description file, but for most cases, this
should not be a problem or can be worked around.

Signed-off-by: Marco Elver <me@marcoelver.com>