pkgix: Add support for plugin commands
Add support for plugin commands. pkgix looks for commands which begin
with "pkgix-", and treats them as additional commands.

There are 2 options for running such a command: 1) source the contents
of the executable file (must be BASH) or 2) run as separate program. In
case 1), all functions available in pkgix are also available to the
plugin command.

Signed-off-by: Marco Elver <me@marcoelver.com>