Add testing-framework and simple test example
This adds a simple testing framework based on:

  1) prefix contents
  2) pkgix stdout/stderr

As none of the actual expected data is stored -- instead checksums are
used -- if discrepancies are found, a previously working version of
pkgix must be used to compare output.

The testing framework itself is a pkgix plugin, and as such has all
available functions from pkgix available to it.

In addition a `check` target was added to the Makefile.

Signed-off-by: Marco Elver <me@marcoelver.com>